Refactoring: weatherData doesnt need to now anything about CSVParsing thats why it will get a DataFactory at construction which encapsulates the data generating from csv

+++ src/main/java/de/exxcellent/challenge/WeatherData.java
@@ -2,18 +2,18 @@ package de.exxcellent.challenge;
 
 public interface WeatherData {
     public int getDay(int index);
-    public int getMxT(int index);
-    public int getMnT(int index);
-    public int getAvT(int index);
+    public float getMxT(int index);
+    public float getMnT(int index);
+    public float getAvT(int index);
     public float getAvDP(int index);
-    public int get1HrP_TPcpn(int index);
-    public int getPDir(int index);
+    public float get1HrP_TPcpn(int index);
+    public float getPDir(int index);
     public float getAvSp(int index);
-    public int getDir(int index);
-    public int getMxS(int index);
+    public float getDir(int index);
+    public float getMxS(int index);
     public float getSkyC(int index);
-    public int getMxR(int index);
-    public int getMn(int index);
+    public float getMxR(int index);
+    public float getMn(int index);
     public float getR_AvSLP(int index);
 
     public int getNumberEntries();
